Refactor to do host interleaving outside the host source
This makes it so hostsources return Host objects that include pool
information and we do the interleaving step at a common level so that
future hostsources can take advantage of it as well.
